final ui/ux

+++ lib/view_reminders_page.dart
@@ -38,20 +38,57 @@ class ViewRemindersPage extends StatelessWidget {
               var reminderName = reminder['name'];
               var reminderDate = reminder['date'].toDate();
 
-              return ListTile(
-                title: Text(reminderName),
-                subtitle:
-                    Text(DateFormat('yyyy-MM-dd – kk:mm').format(reminderDate)),
-                trailing: IconButton(
-                  icon: Icon(Icons.delete),
-                  onPressed: () async {
-                    await FirebaseFirestore.instance
-                        .collection('users')
-                        .doc(userEmail)
-                        .collection('reminders')
-                        .doc(reminder.id)
-                        .delete();
-                  },
+              return Card(
+                elevation: 4.0,
+                margin: EdgeInsets.symmetric(vertical: 8.0, horizontal: 16.0),
+                child: ListTile(
+                  contentPadding:
+                      EdgeInsets.symmetric(vertical: 8.0, horizontal: 16.0),
+                  title: Text(
+                    reminderName,
+                    style: TextStyle(fontSize: 18, fontWeight: FontWeight.bold),
+                  ),
+                  subtitle: Text(
+                    DateFormat('yyyy-MM-dd – kk:mm').format(reminderDate),
+                    style: TextStyle(color: Colors.blueGrey),
+                  ),
+                  trailing: IconButton(
+                    icon: Icon(Icons.delete, color: Colors.red),
+                    onPressed: () async {
+                      bool? confirmDelete = await showDialog(
+                        context: context,
+                        builder: (context) {
+                          return AlertDialog(
+                            title: Text('Confirm Delete'),
+                            content: Text(
+                                'Are you sure you want to delete this reminder?'),
+                            actions: [
+                              TextButton(
+                                onPressed: () =>
+                                    Navigator.of(context).pop(false),
+                                child: Text('Cancel'),
+                              ),
+                              TextButton(
+                                onPressed: () =>
+                                    Navigator.of(context).pop(true),
+                                child: Text('Delete',
+                                    style: TextStyle(color: Colors.red)),
+                              ),
+                            ],
+                          );
+                        },
+                      );
+
+                      if (confirmDelete == true) {
+                        await FirebaseFirestore.instance
+                            .collection('users')
+                            .doc(userEmail)
+                            .collection('reminders')
+                            .doc(reminder.id)
+                            .delete();
+                      }
+                    },
+                  ),
                 ),
               );
             },
